refactor: split module tests into sibling files
- move CLI, command, and event-loop tests into dedicated sibling files
- keep runtime modules focused by removing large inline test blocks
- preserve existing coverage for extracted test suites
